requirements.txt
aiocontextvars==0.2.2
asgiref==3.3.0
autopep8==1.5.7
boto3==1.16.45
botocore==1.19.45
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
contextvars==2.4
coverage==5.3.1
cryptography==3.3.1
decorator==4.4.2
Django==3.1.3
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-filter==2.4.0
django-js-asset==1.2.2
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
djangorestframework==3.12.2
dynaconf==3.1.2
hvac==0.10.6
idna==2.10
immutables==0.16
Jinja2==2.11.2
jmespath==0.10.0
knox==0.1.11
loguru==0.5.3
MarkupSafe==1.1.1
Pillow==8.0.1
pycodestyle==2.7.0
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.4
razorpay==1.2.0
requests==2.25.1
s3transfer==0.3.3
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
validators==0.18.2
win32-setctime==1.0.3
